Add <uses-library> information.
The required <uses-library> "org.apache.http.legacy" is in the manifest
for QuickSearchBox. It is necessary to explicitly add it to `uses_libs`
because it is not among the app dependencies (so Soong does not add it
automatically to the manifest_fixer). It also doesn't add it as a
compatibility library because the target API level for the app is not
below 28.

Bug: 132357300
Test: lunch cf_x86_phone-userdebug && m
Test: m out/soong/.intermediates/packages/apps/QuickSearchBox/QuickSearchBox/android_common/manifest_check/AndroidManifest.xml
Change-Id: I2fe6e09c22b4aa5ef1aac9728c13e5fc829a4468