Fixed getCount() after close() exception in DelayingSuggestionsAdapter
A runtime exception was introduced in DelayingSuggestionsAdapter.kt that remained undetected by unit tests but would crash the app shortly after start. The automatic Kotlin conversion tool converted the getSuggestions() function to a var with an overriden getter. However, the function setPendingSuggestions() also took a parameter of the same name and type, causing this local parameter to be used when getSuggestions() was intended. TODO: file a bug against the Kotlin converter for the case where a class var is created in the conversion with the same name and type as a function parameter.

Bug: 243681952

Test: manual testing with bramble device, unit tests ran with atest QuickSearchBoxTests with result:
Summary (Test executed with 1 devices.)
-------
arm64-v8a QuickSearchBoxTests: Passed: 50, Failed: 0, Ignored: 0, Assumption Failed: 0,

All tests passed!

Change-Id: I5f8349784f8c6b7f1194e0231cec829b123396dd